pyproject.toml
[project]
name = "doctor-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.3.0",
    "gradio>=5.33.0",
    "groq>=0.26.0",
    "gtts>=2.5.4",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "speechrecognition>=3.14.3",
]
